+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.58.5] - 2026-07-04
+
 ### Changed
 
 - **Supervisor Windows launches now default to hidden windows.** `supervisor.json`